It does!
Let's add the functionality for the side bar to collapse and expand so that only the icons are visable. The user could expand the menu out again on clicking the same collapse/expand button.

The size of the side panel would adjust in terms of width accordingly.

If it makes sense, on web and desktop app, the side panel could expand on hover